pyproject.toml
[project]
name = "generative-ai-with-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "duckduckgo-search>=8.1.0",
  "ipykernel>=6.29.5",
  "langchain>=0.3.26",
  "langchain-anthropic>=0.3.15",
  "langchain-community>=0.3.27",
  "langchain-experimental>=0.3.4",
  "langchain-google-genai>=2.1.6",
  "langchain-google-vertexai>=2.0.27",
  "langchain-openai>=0.3.25",
  "langgraph>=0.5.1",
  "python-dotenv>=1.1.1",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "tab"
